hello_world/requirements.txt
aws-lambda-powertools[tracer]
psycopg2-binary
requests
